Signature compact format.
It's just an overly-pretentious regular expression, but heck... it
freaking works, right?

Maybe someday Beeselbob will put CoreParse into CocoaPods, making that a
possibility. CoreParse was a bust for my purposes, but you never know.
Yes, I'd like to move it out of a regex and into a proper BNF grammar.
Thanks for your concern.